pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Hexlet project: Page Analyzer"
authors = [{ name = "Eugene", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "flask",
    "gunicorn",
    "python-dotenv",
    "requests>=2.32.5",
    "psycopg2-binary"
]

[tool.uv]

[dependency-groups]
dev = [
    "flake8>=7.3.0",
]
